[MIG] website_sale_attribute_filter_form_submit: Migration to version 18.0
TT58471
