+++ website_sale_attribute_filter_form_submit/readme/CONFIGURE.md
@@ -1,7 +1,4 @@
-To configure this module, you need to:
+To configure this module, you must do the following:
 
-1.  Go to the frontend /shop url of the website you want to set the
-    manual filters into.
-2.  Hit the Configure menu and enable the option *Manual filters*.
-
-![](../static/description/manual_filter_configuration.png)
+1.  Go to the frontend URL /shop of the website.
+2.  Open the web editor and in the “Customize” tab, enable the *Manual filters* option.
