+++ website_sale_attribute_filter_form_submit/readme/CONTRIBUTORS.md
@@ -5,3 +5,4 @@
 - [Tecnativa](https://www.tecnativa.com):
 
   > - David Vidal
+  > - Pilar Vargas
