+++ website_sale_attribute_filter_form_submit/templates/website_sale.xml
@@ -9,7 +9,7 @@
     >
         <xpath expr="//form[hasclass('js_attributes')]" position="inside">
             <div class='py-2'>
-                <button id="btn-attribute-filter" class="btn btn-primary">
+                <button id="btn-attribute-filter" class="btn btn-secondary w-100">
                     Apply
                 </button>
             </div>
@@ -32,17 +32,20 @@
         >
             <button
                 name="btn_submit_filters_mobile"
-                class="btn btn-primary d-flex py-1 mb-2"
+                class="btn btn-secondary d-flex py-1 mb-2"
                 title="Clear Filters"
             >
                 Apply
             </button>
         </xpath>
         <xpath
-            expr="//div[contains(@t-if,&quot;a.display_type == 'select'&quot;)]"
+            expr="//div[contains(@t-elif,&quot;a.display_type in ('radio', 'pills', 'select', 'multi')&quot;)]"
             position="after"
         >
-            <div class="list-group list-group-flush" t-if="a.display_type == 'select'">
+            <div
+                class="list-group list-group-flush"
+                t-if="a.display_type == 'select' or a.display_type == 'multi'"
+            >
                 <select
                     name="attrib"
                     t-att-id="'%s-%s' % (a.id,v.id)"
@@ -58,7 +61,7 @@
             </div>
         </xpath>
         <xpath
-            expr="//div[contains(@t-if,&quot;a.display_type == 'select'&quot;)]"
+            expr="//div[contains(@t-elif,&quot;a.display_type in ('radio', 'pills', 'select', 'multi')&quot;)]"
             position="attributes"
         >
             <attribute
